Added React-Async-Popup, to begin replacing the homebrew we have going on. Currently delete playlist on place where it is being used.

+++ client/Views/Playlists/components/PlaylistRow.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import useContextMenu from 'react-use-context-menu';
 import { Playlist } from '~logic/Playlist';
 import { Link } from 'react-router-dom';
+import { useConfirm } from 'react-async-popup';
 
 interface PlaylistRowProps {
     playlist: Playlist;
@@ -24,6 +25,10 @@ const PlaylistRow: React.FC<PlaylistRowProps> = (props: PlaylistRowProps) => {
         setCoords(0, 0);
         setVisible(true);
     };
+
+    const [showConfirm] = useConfirm({
+        title: `Are you sure you want to delete ${props.playlist.name}?`
+    });
 
     return (
         <>
@@ -69,9 +74,12 @@ const PlaylistRow: React.FC<PlaylistRowProps> = (props: PlaylistRowProps) => {
                 {props.deletePlaylist && (
                     <div
                         {...bindMenuItems}
-                        onClick={() => {
+                        onClick={async () => {
                             setVisible(false);
-                            props.deletePlaylist(props.playlist.id);
+                            const result = await showConfirm();
+                            if (result) {
+                                props.deletePlaylist(props.playlist.id);
+                            }
                         }}
                     >
                         Delete Playlist
